requirements.txt
aiohttp>=3.8.0          # асинхронные HTTP-запросы
async-timeout>=4.0.0    # таймауты для aiohttp
PyYAML>=6.0             # чтение/запись YAML конфигураций
aiosqlite>=0.18.0       # асинхронная работа с SQLite
lxml>=4.9.0             # парсинг HTML (опционально, если используешь lxml)
beautifulsoup4>=4.12.0  # парсинг HTML (если используешь BS4)
yarl>=1.9.2             # для работы с URL в aiohttp
certifi>=2023.12.0      # SSL-сертификаты для HTTPS
colorlog>=6.7.2023		# используется для красивого цветного логирования
matplotlib>=3.7.2 		# используется для визуализации статистики
requests==2.31.2023		# для тестов
beautifulsoup4==4.12.2 	# для тестов

